rename knn join and distance join algorithm; update example of knn join
