+++ engine/sql/core/src/main/scala/org/apache/spark/sql/execution/joins/BDJSpark.scala
@@ -19,39 +19,37 @@ package org.apache.spark.sql.execution.joins
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions._
+import org.apache.spark.sql.catalyst.util.NumberConverter
 import org.apache.spark.sql.execution.{BinaryNode, SparkPlan}
 import org.apache.spark.sql.partitioner.MapDPartition
 import org.apache.spark.sql.spatial._
-import org.apache.spark.util.BoundedPriorityQueue
 
 import scala.collection.mutable
 import scala.util.Random
 
 /**
   * Created by dong on 1/20/16.
-  * KNN Join based on Nested Loop Approach
+  * Distance Join based on Block Nested Loop Approach
   */
-case class NLKJSpark(left_keys: Seq[Expression],
-                     right_keys: Seq[Expression],
-                     l: Literal,
-                     left: SparkPlan,
-                     right: SparkPlan) extends BinaryNode {
+case class BDJSpark(
+                                   left_keys: Seq[Expression],
+                                   right_keys: Seq[Expression],
+                                   l: Literal,
+                                   left: SparkPlan,
+                                   right: SparkPlan
+                                 ) extends BinaryNode {
   override def output: Seq[Attribute] = left.output ++ right.output
 
   final val num_partitions = sqlContext.conf.numShufflePartitions
-  final val k = l.value.asInstanceOf[Number].intValue()
-
-  private class DisOrdering extends Ordering[(InternalRow, Double)] {
-    override def compare(x : (InternalRow, Double), y: (InternalRow, Double)): Int =
-      -x._2.compare(y._2)
-  }
+  final val r = NumberConverter.literalToDouble(l)
+  final val dimension = left_keys.length
 
   override protected def doExecute(): RDD[InternalRow] = {
     val tot_rdd = left.execute().map((0, _)).union(right.execute().map((1, _)))
 
     val tot_dup_rdd = tot_rdd.flatMap {x =>
       val rand_no = new Random().nextInt(num_partitions)
-      val ans = mutable.ListBuffer[(Int, (Int, InternalRow))]()
+      var ans = mutable.ListBuffer[(Int, (Int, InternalRow))]()
       if (x._1 == 0) {
         val base = rand_no * num_partitions
         for (i <- 0 until num_partitions)
@@ -83,17 +81,17 @@ case class NLKJSpark(left_keys: Seq[Expression],
         }
       }
 
-      val joined_ans = mutable.ListBuffer[(InternalRow, Array[(InternalRow, Double)])]()
+      val joined_ans = mutable.ListBuffer[InternalRow]()
 
-      left_data.foreach(left => {
-        var pq = new BoundedPriorityQueue[(InternalRow, Double)](k)(new DisOrdering)
-        right_data.foreach(right => pq += ((right._2, right._1.minDist(left._1))))
-        joined_ans += ((left._2, pq.toArray))
-      })
+      left_data.foreach {left =>
+        right_data.foreach {right =>
+          if (left._1.minDist(right._1) <= r) {
+            joined_ans += new JoinedRow(left._2, right._2)
+          }
+        }
+      }
+
       joined_ans.iterator
-    }.reduceByKey((left, right) => (left ++ right).sortWith(_._2 < _._2).take(k), num_partitions)
-    .flatMap {
-      now => now._2.map(x => new JoinedRow(now._1, x._1))
     }
   }
 }
