+++ engine/sql/core/src/main/scala/org/apache/spark/sql/execution/joins/BKJSpark.scala
@@ -19,37 +19,39 @@ package org.apache.spark.sql.execution.joins
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions._
-import org.apache.spark.sql.catalyst.util.NumberConverter
 import org.apache.spark.sql.execution.{BinaryNode, SparkPlan}
 import org.apache.spark.sql.partitioner.MapDPartition
 import org.apache.spark.sql.spatial._
+import org.apache.spark.util.BoundedPriorityQueue
 
 import scala.collection.mutable
 import scala.util.Random
 
 /**
   * Created by dong on 1/20/16.
-  * Distance Join based on Nested Loop Approach
+  * KNN Join based on Block Nested Loop Approach
   */
-case class NLDJSpark(
-                                   left_keys: Seq[Expression],
-                                   right_keys: Seq[Expression],
-                                   l: Literal,
-                                   left: SparkPlan,
-                                   right: SparkPlan
-                                 ) extends BinaryNode {
+case class BKJSpark(left_keys: Seq[Expression],
+                    right_keys: Seq[Expression],
+                    l: Literal,
+                    left: SparkPlan,
+                    right: SparkPlan) extends BinaryNode {
   override def output: Seq[Attribute] = left.output ++ right.output
 
   final val num_partitions = sqlContext.conf.numShufflePartitions
-  final val r = NumberConverter.literalToDouble(l)
-  final val dimension = left_keys.length
+  final val k = l.value.asInstanceOf[Number].intValue()
+
+  private class DisOrdering extends Ordering[(InternalRow, Double)] {
+    override def compare(x : (InternalRow, Double), y: (InternalRow, Double)): Int =
+      -x._2.compare(y._2)
+  }
 
   override protected def doExecute(): RDD[InternalRow] = {
     val tot_rdd = left.execute().map((0, _)).union(right.execute().map((1, _)))
 
     val tot_dup_rdd = tot_rdd.flatMap {x =>
       val rand_no = new Random().nextInt(num_partitions)
-      var ans = mutable.ListBuffer[(Int, (Int, InternalRow))]()
+      val ans = mutable.ListBuffer[(Int, (Int, InternalRow))]()
       if (x._1 == 0) {
         val base = rand_no * num_partitions
         for (i <- 0 until num_partitions)
@@ -81,17 +83,17 @@ case class NLDJSpark(
         }
       }
 
-      val joined_ans = mutable.ListBuffer[InternalRow]()
+      val joined_ans = mutable.ListBuffer[(InternalRow, Array[(InternalRow, Double)])]()
 
-      left_data.foreach {left =>
-        right_data.foreach {right =>
-          if (left._1.minDist(right._1) <= r) {
-            joined_ans += new JoinedRow(left._2, right._2)
-          }
-        }
-      }
-
+      left_data.foreach(left => {
+        var pq = new BoundedPriorityQueue[(InternalRow, Double)](k)(new DisOrdering)
+        right_data.foreach(right => pq += ((right._2, right._1.minDist(left._1))))
+        joined_ans += ((left._2, pq.toArray))
+      })
       joined_ans.iterator
+    }.reduceByKey((left, right) => (left ++ right).sortWith(_._2 < _._2).take(k), num_partitions)
+    .flatMap {
+      now => now._2.map(x => new JoinedRow(now._1, x._1))
     }
   }
 }
